move `ty::List` into a new submodule
